Move output_settings to scripts/Engineering/common
