+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/Engineering/gui/engineering_diffraction/tabs/focus/model.py
@@ -7,7 +7,7 @@
 from typing import Optional
 
 import Engineering.EnggUtils as EnggUtils
-from mantidqtinterfaces.Engineering.gui.engineering_diffraction.tabs.common import output_settings
+from Engineering.common import output_settings
 from Engineering.common.calibration_info  import CalibrationInfo
